Test for part1 done, seems to work

+++ hw5/include/map_reduce.h
@@ -2,7 +2,7 @@
 #define _MAP_REDUCE_H_
 
 struct map_res{
-	size_t datum_cnt;
+	unsigned long datum_cnt;
 	char *filename;
 	unsigned long tot_duration; // total duaration
 	struct list *year_root;
